+++ ClearKeep/App/Resources/Localize.strings
@@ -67,9 +67,6 @@
 "Login.ForgotPassword" = "Forgot Password?";
 "Login.SignUp.Suggest" = "Don't have an account?";
 "Login.SignUp" = "Sign up";
-"General.DisplayName" = "Display Name";
-"General.Password" = "Password";
-"General.ConfirmPassword" = "Confirm Password";
 
 // MARK: - Register
 "Register.Title" = "Please fill in the information below to complete your sign up";
@@ -77,12 +74,14 @@
 "Register.SignUp" = "Sign up";
 
 // MARK: - ForgotPassword
-"ForgotPass.TitleMail" = "Please enter your email to reset your password";
-"ForgotPass.TitleChangePassword" = "Please enter your details to change password";
-"ForgotPass.Forgotpassword" = "Forgot password";
-"ForgotPass.NewPassword" = "Enter Your New Password";
-"ForgotPass.Resetpassword" = "Reset password";
-"ForgotPass.Save" = "Save";
+"ForgotPassword.Title" = "Forgot password";
+"ForgotPassword.Description" = "Please enter your email to reset your password";
+"ForgotPassword.ResetPassword" = "Reset password";
+
+// MARK: - NewPassword
+"NewPassword.Title" = "Enter Your New Password";
+"NewPassword.Description" = "Please enter your details to change password";
+"NewPassword.NewPassword" = "New Password";
 
 // MARK: - AdvancedServer
 "AdvancedServer.Title" = "Please enter your server URL to enter custom server";
